Refactor crypto.rs into secp256k1.rs and ed25519.rs
