+++ packages/crypto/src/lib.rs
@@ -4,15 +4,18 @@
 //! This crate does not adhere to semantic versioning.
 #![cfg_attr(feature = "backtraces", feature(backtrace))]
 
-mod crypto;
+mod ed25519;
 mod errors;
 mod identity_digest;
+mod secp256k1;
 
 #[doc(hidden)]
-pub use crate::crypto::{ed25519_verify, secp256k1_verify};
+pub use crate::ed25519::ed25519_verify;
 #[doc(hidden)]
-pub use crate::crypto::{ECDSA_PUBKEY_MAX_LEN, ECDSA_SIGNATURE_LEN, MESSAGE_HASH_MAX_LEN};
-#[doc(hidden)]
-pub use crate::crypto::{EDDSA_PUBKEY_LEN, EDDSA_SIGNATURE_LEN, MESSAGE_MAX_LEN};
+pub use crate::ed25519::{EDDSA_PUBKEY_LEN, EDDSA_SIGNATURE_LEN, MESSAGE_MAX_LEN};
 #[doc(hidden)]
 pub use crate::errors::{CryptoError, CryptoResult};
+#[doc(hidden)]
+pub use crate::secp256k1::secp256k1_verify;
+#[doc(hidden)]
+pub use crate::secp256k1::{ECDSA_PUBKEY_MAX_LEN, ECDSA_SIGNATURE_LEN, MESSAGE_HASH_MAX_LEN};
